Added the excel document for maximum accessibility
The excel document contains a set of tables which have been linked using the Data/Relationships dialogue. It has a series of sheets to sepparate the data from the displays. The sheets are: Display (for graphs); Tables (for the pivot tables the graphs are based on); Data (the tables that contain the data); Original Data (the non-relational version of a portion of the data. Redundant. Can be deleted)