Uploaded SQL files for rebuilding the database
The SQL files are not set up as relational but you can create the links and views as you see fit. It's all named the same as the Excel document so if in doubt, check how that's set up.

SQL is more versatile than Excel so if you want to evaluate the data in more interesting ways, use SQL views.